refactor(model): fold tokens + login_attempts into the composition root (#2858)
